requirements.txt
tensorflow-cpu==2.11.0
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
tensorboard==2.11.0
matplotlib==3.6.3
streamlit==1.26.0
